split remaining map suites
